parser_tests/requirements-parsers.txt
pypdf==4.1.0
pdfplumber==0.11.0
PyMuPDF>=1.24.2
pymupdf4llm==0.0.11
llama-parse
python-dotenv
azure-ai-documentintelligence
pytesseract
pdf2image
